% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{addMAE}
\alias{addMAE}
\title{Calculate the median absolute error over categories and add to a plot}
\usage{
addMAE(data, y, x, gg = NULL, nsmall = 3)
}
\arguments{
\item{data}{The data used in the plot}

\item{y}{A character vector of length 2 that specifies the
columns to summarize}

\item{x}{A character value providing the right side of the
formula used for aggregation (e.g., \code{"depth + year"})}

\item{gg}{The \code{\link{ggplot2}} you want to add the MAE to}

\item{nsmall}{An integer value providing the number of digits you
want after the decimal for the MAE}

\item{importFrom}{stats aggregate median}
}
\description{
Calculate the median absolute error over categories and add to a plot
}
