% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rtools.R
\name{check_rtools}
\alias{check_rtools}
\title{Check Rtools is installed properly}
\usage{
check_rtools()
}
\value{
Various summaries are printed to the screen to help you troubleshoot
your installation of Rtools. If you make it to the end of this function
successfully, then a value of 0 will be returned and you should be
able to then run \code{\link[=check_TMB]{check_TMB()}} to see if TMB is properly
installed for your system.
}
\description{
Check that Rtools is properly installed on your machine such that
packages and code within packages can be compiled.
}
\details{
Warning - this function was built by Windows users and may not
be sufficient for other operating systems.

A new version of Rtools is required for R releases > 4.0.0
https://cran.r-project.org/bin/windows/Rtools/
}
\examples{
# Hope for a 0
check_rtools()
}
\seealso{
\code{\link[=check_TMB]{check_TMB()}} to see if TMB is properly installed
and has access to rtools.
}
\author{
Chantel R. Wetzel, modified by Kelli F. Johnson
}
