% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_TMB.R
\name{check_TMB}
\alias{check_TMB}
\title{Check TMB is working}
\usage{
check_TMB()
}
\value{
A single integer is returned, where zero means that
your installation of rtools40 and TMB is good to go. I hope you get a zero!
If the status is failed, then you will see \code{"Compilation failed"},
a stop message generated by \code{\link[TMB:compile]{TMB::compile()}}.
}
\description{
Check that the installation of TMB is working before trying
to run any VAST models.
}
\details{
This function will generate a lot of output to your screen that
can be helpful when trying to troubleshoot the installation of
rtools40. Most importantly, you can see the location of where
your computer is trying to find make files. If you have already run this
function on your computer, then you should see
\verb{make: Nothing to be done for 'all'}.
If the installation of TMB has failed, then you will see
\code{"Error in TMB::compile(...)"}.
The most likely culprit for it failing is a faulty installation of
rtools, specifically rtools40 for R version 4.0 and greater.
}
\examples{
# Hope for a 0
check_TMB()
# If you get something other than 0, run
check_rtools()

}
\seealso{
\link{check_rtools} for checking the installation of Rtools.
}
\author{
Kelli Faye Johnson
}
