% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Database for VAST}
\usage{
clean_data(data)
}
\arguments{
\item{data}{A data frame that you want to use for a 
VAST model.}
}
\description{
Many specific column names are needed for VAST but the data
are downloaded from several repositories, thus making cleaning
the data a necessary step before running VAST. This function 
attempts to standardize the columns to work with VAST.
}
\details{
todo: figure out area swept for the different surveys
WCGBTS and all AFSC surveys are in KG/Hectare but early on
I used "AreaSwept_km2" = 0.01
todo: ensure that there are no \code{NA} values, i.e., na.omit(data)
}
\author{
Kelli Faye Johnson
}
