% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unload.R
\name{clean_unload}
\alias{clean_unload}
\title{Unload VAST dll}
\usage{
clean_unload(searchfor = "VAST")
}
\arguments{
\item{searchfor}{A file name of the dll to search for.}
}
\value{
The name(s) of the unloaded libraries are returned.
}
\description{
Unload a dll library with the VAST name.
}
\author{
Kelli Faye Johnson
}
