% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_strata4plot.R
\name{convert_strata4plot}
\alias{convert_strata4plot}
\title{Convert Strata Names for Plot Text}
\usage{
convert_strata4plot(text)
}
\arguments{
\item{text}{A vector of text strings that will be formatted.}
}
\value{
A vector of text strings the same length as \code{text} that
have been formatted to use in a plot or in text.
}
\description{
Convert the strata names in the column \code{STRATA} to names that
are formatted for a figure or for text. Things such as converting
underscores to spaces and two-letter lower-case words to two-letter
upper-case words are performed for you.
}
\author{
Kelli Faye Johnson
}
