% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_strata4vast.R
\name{convert_strata4vast}
\alias{convert_strata4vast}
\title{Convert Area-Based Strata to VAST Strata}
\usage{
convert_strata4vast(
  strata,
  overridedepth = FALSE,
  mindepth = 55,
  maxdepth = 1280
)
}
\arguments{
\item{strata}{A data frame generated using
\code{\link[nwfscSurvey]{GetStrata.fn}()}.}

\item{overridedepth}{A logical value specifying if \code{mindepth} and
\code{maxdepth} should be used instead of the depths that are available
within \code{strata}. These depths will always replace NA values, but
if \code{TRUE} all depths will be replaced with the argument values.
The default is to not replace them, i.e., \code{overridedepth = FALSE}.}

\item{mindepth}{The depth in meters that you want to use as a boundary for
the shallow edge of all strata. The default value is 55 m.
All \code{NA} values will be replaced with \code{mindepth}.}

\item{maxdepth}{The maximum depth that you want to use as a boundary for
the deep edge of all strata. The default value is 1280 m.
All \code{NA} values will be replaced with \code{maxdepth}.}
}
\value{
A data frame with one row per strata.
Strata can overlap and need not be unique.
\itemize{
\item STRATA (the name of the strata, only used for plotting later),
\item north_border (northern latitude of the strata),
\item south_border (southern latitude of the strata),
\item shallow_border (shallow depth (m) of the strata), and
\item deep_border (deep depth (m) of the strata).
}
}
\description{
Convert strata defined using \code{\link[nwfscSurvey]{GetStrata.fn}()}
to strata that will work with VAST. Results will be more consistent using
these defined strata rather than All_Areas, which predicts to the entire
extrapolation grid.
}
\examples{
convert_strata4vast(nwfscSurvey::GetStrata.fn("sablefish"))

}
\author{
Kelli Faye Johnson
}
