% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_survey4vast.R
\name{convert_survey4vast}
\alias{convert_survey4vast}
\title{Convert Survey Name to VAST Column Name}
\usage{
convert_survey4vast(survey)
}
\arguments{
\item{survey}{A character value supplying the survey name.
The value is typically an acronym.
See \code{\link[nwfscSurvey]{createMatrix}()} for a list of defaults.}
}
\value{
A vector of character values, one for each input value, that
specifies the column of the grid data to use for the area sampled.
The default is to return the NWFSC WCGBTS name, thus even values that
don't pertain to a survey will return \code{"propInWCGBTS"}.
}
\description{
Convert a character string that specifies the survey name in terms
of \pkg{nwfscSurvey} to that used in the \code{\link{get_inputgrid}}
column of the grid provided by \pkg{VAST}.
}
\examples{
test <- convert_survey4vast(c("Triennial", "NWFSC.Combo"))
\dontshow{
testthat::expect_true(all(c("propInTriennial", "propInWCGBTS") == test))
}
}
\author{
Kelli Faye Johnson
}
