% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\docType{data}
\name{Database}
\alias{Database}
\title{Example Database from 2019 for Sablefish}
\description{
An example data base pull that was used for the 2019 sablefish stock
assessment. This data is used for testing purposes and to demonstrate
examples. The data set was generated using
\code{\link[nwfscSurvey](PullCatch.fn)}.
}
\examples{
data("Database", package = "VASTWestCoast")
table(Database[, "Year"])

}
\keyword{data}
