% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficient.R
\name{get_coefficient}
\alias{get_coefficient}
\title{Calculate the delta coefficient of a model}
\usage{
get_coefficient(data)
}
\arguments{
\item{data}{A data frame with columns for each year of the index
generated for the operating model and estimated from the estimation
method.}
}
\value{
A vector of delta values, one for each row of the data frame.
}
\description{
Calculate the delta coefficient of a simulation. The coefficient
represents the relationship between the estimation method and the
operating model log-index of abundance. The coefficient should be
close to one, which indicates that the index is well calibrated.
}
\examples{
fake <- data.frame("om_index1" = 2, "om_index2" = 4, "om_index3" = 4,
  "em_index1" = 10, "em_index2" = 18, "em_index3" = 18)
get_coefficient(fake)

}
\author{
Kelli Faye Johnson
}
