% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Obtain Data for a VAST Model}
\usage{
get_data(survey, species, data = NULL)
}
\arguments{
\item{survey}{A character value supplying the survey name.
The value is typically an acronym.}

\item{species}{A character value supplying the scientific name.
The value should be spelled correctly and split using a space.}

\item{data}{A data frame to be cleaned rather than pulling
information from a warehouse.}
}
\description{
Get and clean data to use in a VAST model. The information
can be pulled from a data warehouse or supplied using the
\code{data} argument.
}
