% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate.R
\name{get_depth}
\alias{get_depth}
\title{Get depth from downloaded data}
\usage{
get_depth(a, b, n, squared = TRUE)
}
\description{
Get depth from downloaded data
}
\examples{
 exdepth <- get_depth(Extrapolation_List$Data_Extrap,
   Spatial_List$PolygonList$NN_Extrap$nn.idx, max(Data_Geostat[, "Year"]))
}
