% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depthbin.R
\name{get_depthbin}
\alias{get_depthbin}
\title{Get Depth by Year}
\usage{
get_depthbin(df, vline = 366)
}
\arguments{
\item{df}{A data frame from the NWFSC warehouse.}

\item{vline}{The x-axis location of the vertical line.
The default is to include a line at 366 m because of the
change in sampled depths in the Triennial Survey. A value of
\code{vline = NULL} will remove the line from the plot.}
}
\value{
A ggplot object.
}
\description{
Get depth of positive and empty tows by year plotted in a histogram.
}
\examples{
get_depthbin(Database)
\dontrun{
tridata <- nwfscSurvey::PullCatch.fn(SurveyName = "Triennial")
tridata[, "Catch_KG"] <- tridata[, "total_catch_wt_kg"]
for (ii in unique(tri$Common_name)) {
  dd <- subset(tri, Common_name == ii)
  if (nrow(dd) == 0) next
  g <- get_depthbin(dd)
  g <- g + ggtitle(ii)
  ggsave(g, file = paste0(tempdir(), "Tri_", ii, ".png"))
}
}

}
\author{
Kelli Faye Johnson
}
