% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EM.R
\name{get_EM}
\alias{get_EM}
\title{Get information on parameters from an estimation folder}
\usage{
get_EM(file)
}
\arguments{
\item{file}{An R file with information saved from \code{\link{VAST_EM}}.}
}
\value{
A data frame with parameter values. Parameter names will be
prefixed with "em".
}
\description{
Get information on parameters from an estimation folder
}
\author{
Kelli Faye Johnson
}
