% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EM.R
\name{get_EMpath}
\alias{get_EMpath}
\title{Create full path to estimation model}
\usage{
get_EMpath(simnum, omnum, repnum, emnum, basedir = getwd(),
  simname = "_VAST_simulation", check = FALSE, make = FALSE)
}
\arguments{
\item{basedir}{A file path of your base directory}
}
\description{
Create full path to estimation model
}
\examples{
get_EMpath(simnum = 1, omnum = 1, repnum = 1, emnum = 1)
}
