% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extraplist.R
\name{get_extraplist}
\alias{get_extraplist}
\title{Get Extrapolation List to Expand Results}
\usage{
get_extraplist(survey, strata = NULL)
}
\arguments{
\item{survey}{A character value supplying the survey name.
The value is typically an acronym.}

\item{strata}{A data frame specifying the strata.}
}
\description{
Use \code{\link[FishStatsUtils]{make_extrapolation_info}}
sub functions to prepare the extrapolation grid.
}
\author{
Kelli Faye Johnson
}
