% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputgrid.R
\name{get_inputgrid}
\alias{get_inputgrid}
\title{Get the Input Grid for a Given Survey}
\usage{
get_inputgrid(survey)
}
\arguments{
\item{survey}{A character value supplying the survey name.
The value is typically an acronym.
See \code{\link[nwfscSurvey]{createMatrix}()} for a list of defaults.}
}
\description{
Define the input grid for a given survey, where positive values for
each coordinate pair define the area in square kilometers to predict to.
}
\author{
Kelli Faye Johnson
}
