% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logratio.R
\name{get_logratio}
\alias{get_logratio}
\title{Calculate the log-ratio of the index data.}
\usage{
get_logratio(data)
}
\arguments{
\item{data}{A data frame or vector of index data. Data should be
on the normal scale and will be transformed to the log scale.}
}
\value{
A vector of values log-index values.
}
\description{
Calculate the log-ratio of the index data.
}
\examples{
data <- data.frame("first" = rlnorm(10), "second" = rlnorm(10))
ratios <- get_logratio(data)

}
\author{
Kelli Faye Johnson
}
