% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{get_long}
\alias{get_long}
\title{Wrapper for Gather function
A wrapper function that plays a limited roll in this
package.}
\usage{
get_long(data, names, item, value)
}
\arguments{
\item{data}{A data frame}

\item{names}{A vector of names}

\item{item}{The key}

\item{value}{The value
todo: document better}
}
\description{
Wrapper for Gather function
A wrapper function that plays a limited roll in this
package.
}
