% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mesh.R
\name{get_mesh}
\alias{get_mesh}
\title{Get Mesh for West Coast}
\usage{
get_mesh(data.inner, data.outer, numknots)
}
\arguments{
\item{data.inner}{A data frame of points with at least two columns,
\code{"Lon"} specifying the Longitude and
\code{"Lat"} specifying the Latitude.
The locations will be used to define the inner boundary of the
triangulation mesh and starting seeds for the knot locations.}

\item{data.outer}{A data frame of points with at least two columns,
\code{"Lon"} specifying the Longitude and
\code{"Lat"} specifying the Latitude.
The locations will be used to define the outer boundary of the
triangulation mesh.}

\item{numknots}{The desired number of knots to have in the resulting mesh.
This will be approximate because the distance between points used for knots
will be used to solve to the ideal number of knots using a while loop.
Thus, expect the resulting number of knots to be within plus or minus five
percent of this input value. See \code{\link{get_mesh}()} for specifics.
The desired value is typically stored in the settings list; see,
\code{\link{get_settings}()} and is known as \code{"n_x"} in
\code{\link[FishStatsUtils]{make_settings}()}.}
}
\value{
An \code{inla.mesh} list object created with
\code{\link[INLA]{inla.mesh.create}()}.
The resulting list has 8 default elements from INLA and a few additional
elements that are added specifically for VASTWestCoast, including the
coordinates used to create the mesh. Of most interest is probably the
meta data, which is stored in \code{"meta"} and includes the call to
\code{\link[INLA]{inla.mesh.create}()} and the \code{fmesher.args} used
internally. The call to \code{get_mesh()} is also stored in \code{"call"}.

Users can compare the desired number of knots \code{n.desired} to the
actual number of knots \code{n}. The number of knots should always be within
plus or minus five percent of desired number unless a solution is not
found within 100 tries, which is the current cap on the number of times
the code will run through the \code{while} loop.
}
\description{
Create a mesh for the US West Coast. This method provides
a work around for needing a kmeans file.
}
\author{
Kelli Faye Johnson
}
