% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mesh.R
\name{get_mesh.tmb}
\alias{get_mesh.tmb}
\title{Get Mesh Using Settings from TMB Example}
\usage{
get_mesh.tmb(loc)
}
\arguments{
\item{loc}{A matrix of locations, where the first column is the x
coordinate and the second column is the y coordinate.
The units of the locations are uncertain, and thus, the units of the
specifications used when creating the mesh are also uncertain and will
not be universal.}
}
\description{
Get a triangulation mesh based on the same settings that are used for the
[example](https://github.com/kaskr/adcomp/blob/master/tmb_examples/spde_mesh.R)
in Template Model Builder (TMB).
This function acts as a way to offer documentation and comparison to other
ways to create meshes and is not actually used within the package.
}
