% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OM.R
\name{get_OM}
\alias{get_OM}
\title{Get information on parameters from a simulated data set}
\usage{
get_OM(file)
}
\arguments{
\item{file}{An R file with information saved from \code{\link{VAST_OM}}.}
}
\value{
A data frame with parameter values. Parameter names will be
prefixed with "om".
}
\description{
The function can be ran on multiple operating iterations using
\code{dir(folderx, )}
}
\author{
Kelli Faye Johnson
}
