% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Get results from OM and EM from a simulation}
\usage{
get_results(omdir, write2disk = TRUE)
}
\arguments{
\item{omdir}{The directory that houses the operating model and runs.}

\item{write2disk}{A logical value specifying whether or not to write
the results to the disk. If \code{"results.csv"} already exists one
directory up from the \code{omdir} the results will be appended to this
file.}
}
\value{
A data frame of results with one row per simulation replicate.
}
\description{
Get results from OM and EM from a simulation
}
\author{
Kelli Faye Johnson
}
