% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings_single}
\alias{get_settings_single}
\title{Get Settings For a Single Parameter}
\usage{
get_settings_single(par, default)
}
\arguments{
\item{par}{A vector of values that do not need to be named.
If a single value is provided, it will be repeated for each needed value.}

\item{default}{A vector of default values that will be used if
\code{par = NULL}. The names from this vector are always copied
over to the vector that is returned, so name them well.}
}
\value{
A vector of values the same length as \code{default} and with the
same names as those given to \code{default}.
}
\description{
Get Settings For a Single Parameter
}
