% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_sim}
\alias{get_sim}
\title{Simulate data}
\usage{
get_sim(Sim_Settings = NULL, Extrapolation_List, Data_Geostat = NULL,
  standardize_fields = FALSE)
}
\arguments{
\item{Sim_Settings}{An optional tagged list for specifying input parameters. 
Users can look in the code to see the detailed components of the list and what
the missing values will be set to. If \code{NULL}, then all parameters will be set
to their default values within the function. Values used will be returned in the 
output list under the argument name.}

\item{Extrapolation_List}{Output from \code{\link[FishStatsUtils]{make_extrapolation_info}}}

\item{Data_Geostat}{A data frame with column headers 
\code{c("Lon", "Lat", "Year", "Vessel", "AreaSwept_km2")} 
containing the sample design to mimic.}

\item{standardize_fields}{Boolean, whether to ensure that random fields have sample mean and standard deviation equal to their inputted values}
}
\value{
Return Tagged list of output
\describe{
  \item{Data_Geostat}{Simulated data for analysis}
  \item{B_tl}{True biomass for each year and stratum}
}
}
\description{
Simulates data for testing \code{VAST}.
The original function (SpatialDeltaGLMM::Geostat_Sim) was available in 
devtools::install_github("nwfsc-assess/geostatistical_delta-GLMM",
  ref = "da3a3badfd0c50cc805fc5439149fb3370c3791e")
}
\examples{
## # Prepare inputs
## Extrapolation_List <- FishStatsUtils::make_extrapolation_info(
## Region = "Eastern_Bering_Sea", strata.limits=data.frame('STRATA'="All_areas"))
## test <- get_sim(Extrapolation_List = Extrapolation_List)

}
