% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp.R
\name{get_spp}
\alias{get_spp}
\title{Get the Species and Survey Name}
\usage{
get_spp(input, split = "_")
}
\arguments{
\item{input}{A character value that contains the survey of
interest and the scientific name of the species. The three
words will be split using the \code{split} argument.}

\item{split}{A character value specifying how to break up
the \code{input} argument. The default is to use \code{"_"}.}
}
\value{
A vector with two entries, survey specifies the survey
name and species provides a normalized scientific name.
}
\description{
Separate a single character string into three words to 
determine the survey and scientific name for the species
of interest.
}
\examples{
get_spp(input = "Triennial_Anoplopoma_fimbria")

}
\author{
Kelli Faye Johnson
}
