% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{ggplotre}
\alias{ggplotre}
\title{Wrapper for ggplot2 with Relative Error
A wrapper for printing ggplots that display relative
error.
todo: document better}
\usage{
ggplotre(data, x, y, print = TRUE, gradient = FALSE,
  facetx = c("emname", "em_depth"), facety = "omname", labels = NULL,
  type = "box", scales = c("fixed", "free"), dir = NULL,
  lim = list(x = NULL, y = NULL))
}
\arguments{
\item{data}{A data frame}

\item{x}{The x variable}

\item{y}{The y variable}

\item{print}{A logical}

\item{gradient}{A logical suggesting if the gradient should
be displayed}

\item{facetx}{A character vector for x facets}

\item{facety}{A character vector for y facets}

\item{labels}{Labels for the x and y axes.}

\item{type}{What type of plot you want, e.g., code{"box"}}

\item{scales}{If you want the scales fixed or free}

\item{dir}{A file path where you want the plot saved}

\item{lim}{A list of x and y limits}
}
\description{
Wrapper for ggplot2 with Relative Error
A wrapper for printing ggplots that display relative
error.
todo: document better
}
