% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.inputgrid.R
\name{plot.inputgrid}
\alias{plot.inputgrid}
\title{Plot input grid}
\usage{
\method{plot}{inputgrid}(data, savedir = getwd(), lab.survey = "survey")
}
\arguments{
\item{data}{A data frame returned from \link{get_inputgrid}
that includes the following columns:
\itemize{
\item Latitude, available as \code{"Lat"};
\item Longitude, available as \code{"Lon"};
\item Area, available as \code{"Area_km2"}; and
\item Depth in meters.
}}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{lab.survey}{A character entry specifying the name of the survey that
gathered the data. This can be text of your choice or a single word used
to represent the survey name as defined in the first column of
\code{\link[nwfscSurvey]{createMatrix}()}.
See \code{\link[nwfscSurvey]{GetSurveyAbb.fn}()}
for how this shortened name is linked to a longer abbreviation.
Either the results of this function
based on your input or your direct input if no match is found will be
used as text on the plot or in the resulting caption.
The default is the word \code{"survey"} which will fit the bill for a
generic specification.}
}
\value{
A figure 9 inches long (roughly fitting on a single page)
is saved to the disk, along with a caption. The files are saved
using the name VASTWestCoast_mesh.xxx. The \pkg{ggplot2} object
is returned for future manipulation.
}
\description{
Plot a map of the grid cells used in the prediction of VAST results.
}
\author{
Kelli Faye Johnson
}
