% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mesh.R
\name{plot.mesh}
\alias{plot.mesh}
\title{Plot mesh}
\usage{
\method{plot}{mesh}(mesh, savedir = getwd(), lab.survey = "survey")
}
\arguments{
\item{mesh}{An \pkg{INLA} mesh resulting from \code{\link[=get_mesh]{get_mesh()}}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{lab.survey}{A character entry specifying the name of the survey that
gathered the data. This can be text of your choice or a single word used
to represent the survey name as defined in the first column of
\code{\link[nwfscSurvey]{createMatrix}()}.
See \code{\link[nwfscSurvey]{GetSurveyAbb.fn}()}
for how this shortened name is linked to a longer abbreviation.
Either the results of this function
based on your input or your direct input if no match is found will be
used as text on the plot or in the resulting caption.
The default is the word \code{"survey"} which will fit the bill for a
generic specification.}
}
\value{
A figure nine inches long (roughly fitting on a single page)
is saved to the disk, along with a caption. The files are saved
using the name VASTWestCoast_mesh.xxx. The \pkg{ggplot2} is also
returned for future manipulation.
}
\description{
Plot a mesh created from the \link{get_mesh} that is saved in
the specified directory as \code{VASTWestCoast_mesh}.
}
\seealso{
See \code{\link[INLA:plot.inla.mesh]{INLA::plot.inla.mesh()}} for regular base R
methods for plotting meshes, which can be called using
\code{plot(meshobject)}.
}
\author{
Kelli F. Johnson
}
