% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caption.R
\name{plot_caption}
\alias{plot_caption}
\title{Make and Save a Figure Caption}
\usage{
plot_caption(text, survey, figname, savefile)
}
\arguments{
\item{text}{A text string of any length pasted together into a single string.}

\item{survey}{A character value supplying the survey name.
The value is typically an acronym.
See \code{\link[nwfscSurvey]{createMatrix}()} for a list of defaults.}

\item{figname}{A character value that will be pasted together along with the
survey name to generate the label used in the Latex file or markdown file.
For example entries of \code{figname = "mesh"} and \code{survey = "Triennial"}
would lead to a label of \code{"fig_mesh_Triennial"}.}

\item{savefile}{The file name that you want the results saved to. A full path can
be specified, else the file will be saved in the current directory.}
}
\value{
The resulting data frame with the label and the caption as column
entries with a single row for the given figure. The data frame is also
saved to the disk using the file name in \code{savefile}.
}
\description{
Take input text and generate a standardized csv file
with a caption for a given figure.
}
