% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot indicies of Relative Abundance}
\usage{
plot_index(
  dir,
  recursive = TRUE,
  area = "All_areas",
  keepyears = NULL,
  limit = NULL
)
}
\arguments{
\item{dir}{A file path with folders containing VASTWestCoast runs.}

\item{recursive}{A boolean value indicating whether or not you want to
recursively search for VAST runs inside \code{dir} or just look 
at the top level.}

\item{area}{A character value specifying the area to plot.
The area is specified as the fleet column in the csv file.}

\item{keepyears}{A vector of years you want to plot.
If \code{NULL}, then all years will be plotted.}

\item{limit}{A vertical-axis upper limit, which is forced.}
}
\description{
Compare estimated indices by plotting them with their confidence
intervals on a single plot.
}
