% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_histogram.R
\name{plot_multi_histogram}
\alias{plot_multi_histogram}
\title{Plot Multiple Histograms}
\usage{
plot_multi_histogram(df, col.feature, col.label)
}
\arguments{
\item{df}{A data frame.}

\item{col.feature}{A character value specifying the column of \code{df}
that will be used for the x axis.}

\item{col.label}{A character value specifying the column of \code{df}
that will be used for the colors or indivdual lines.}
}
\value{
A ggplot object.
}
\description{
Plot multiple histograms for a given column (x axis) by
category (colors). Both the stacked histogram and individual
density lines are included, where the colors coincide and are
labelled in the margin.
}
\author{
Kelli F. Johnson
}
