% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot quantiles versus theoretical quantiles}
\usage{
plot_qq(outdir, model)
}
\arguments{
\item{outdir}{The directory where you would like to save the resulting figure,
which will be named \code{VASTWestCoast_QQ.png}.}

\item{model}{The list of model specifications and results, typically available
in the \code{Save.RData} file and produced by \link{VAST_do}.}
}
\description{
Plot quantiles versus theoretical quantiles as was originally done in \link{VAST}
prior to their use of DHARMa residual diagnostics.
Functionality is limited to lognormal and gamma distributions for a single
species or group.
The 1:1 line is automatically added to the figure to aid in performing the
fat-pen test to determine if your distributional assumption fits your data.
}
\seealso{
Users will not typically interact with this function, but will instead
benefit from it being called in \link{VAST_diagnostics}.
}
\author{
Kelli Faye Johnson
}
