% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ss.R
\name{plot_ss}
\alias{plot_ss}
\title{Plot Index Time Series}
\usage{
plot_ss(file.in = "Table_for_SS3.csv", file.out, name = "", do_smooth = TRUE)
}
\arguments{
\item{file.in}{A file path pointing to the \code{"Table_for_SS3.csv"} file
that has the results you wish to plot. The path can be relative or absolute.
This file is generated from \code{\link{VAST_condition}}.}

\item{file.out}{A file path for plotting the results and saving them to the
disk. This can be a relative or absolute path. If no extension is added, then
\code{.png} will be added to the file name because the resulting figure is
a png. If missing, a default file name will be generated.}

\item{name}{A character string that will be pasted after the survey name to
for the main title for the figure.
Typically, this is the name of the species, which is also the folder
in which the results exist, though not always.}
}
\value{
A 10 by 7 inch png is saved to the disk showing a time series of the
estimated index of abundance from a single model run. If there are more than
one strata, than all strata are plotted, but the summary strata is always
plotted in black and will be the only line with confidence intervals.
The survey name and text provided in the name argument are pasted together
to form the main title of the plot. Estimates are in metric tonnes and the
ylab reflects this. Years with estimates of zero are excluded from the figure.
A caption with a Latex label is also written to a csv file, where the first
column is the label and the second column is the caption.
}
\description{
Plot a single time series of estimated index values for the
US West Coast.
}
\author{
Chantel Wetzel
}
