% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_greek.R
\name{show_greek}
\alias{show_greek}
\title{Show list of greek letters and their meaning}
\usage{
show_greek(type = c("VAST"), plot = FALSE)
}
\arguments{
\item{type}{The model type that you want to get the symbols for.
See the function call for the available types where the first one
listed is the default using \link{match.arg}.}

\item{plot}{A logical value specifying if a plot should be generated.
The default is to NOT plot the results.}
}
\value{
A data frame with two columns, the greek symbol in words and
the description of what it means in the modeling framework.
}
\description{
Show a list of greek letters and their meaning within the VAST modeling
framework because it is often hard to remember what part of the model
each letter represents.
}
\author{
Kelli Faye Johnson
}
