% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nwfsc.R
\name{summary_nwfsc}
\alias{summary_nwfsc}
\title{Summarize a VAST model for Northwest Fisheries Science Center}
\usage{
summary_nwfsc(obj, parameter_estimates, savedir = NULL)
}
\arguments{
\item{obj}{A compiled object from Template Model Builder (TMB)
after estimating the maximum likelihood estimate of the fixed effects.}

\item{parameter_estimates}{Output from \code{\link[FishStatsUtils]{fit_model}}.}

\item{savedir}{A directory to save the resulting tables, which will
be in csv format.}
}
\value{
Tagged list containing objects for running a VAST model
\describe{
\item{TableA}{Table of settings}
\item{TableB}{Table of estimated parameters}
\item{TableC}{Table of MLE of fixed effects}
}
}
\description{
Generate output tables requested by the Pacific Fisheries
Management Council Science and Statistical Committee
}
\author{
James T. Thorson
}
