% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_session.R
\name{summary_session}
\alias{summary_session}
\title{Save Detailed Session Info to the Disk}
\usage{
summary_session(savefile = "session.txt")
}
\arguments{
\item{savefile}{The file name that you want the results saved to. A full path can
be specified, else the file will be saved in the current directory.}
}
\value{
A list with information about the current R session. Also,
the same list is written to a text file and an \code{RData} file.
}
\description{
Save details about the R session used to run the model to the disk.
Information regarding packages and their versions is saved to
help those trying to recreate results in the future.
}
\examples{
info <- summary_session()
\dontrun{
names(info)
}

}
\author{
Kelli Faye Johnson
}
