% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_diagnostics.R
\name{VAST_diagnostics}
\alias{VAST_diagnostics}
\title{Run Diagnostic Plots for VASTWestCoast Output}
\usage{
VAST_diagnostics(dir = getwd())
}
\arguments{
\item{dir}{The directory where the output from 
\code{\link{VAST_condition}} is saved. Can be a relative
or an absolute path.}
}
\value{
todo: Make a list of the tables and figures that are 
saved to the disk from VAST_diagnostics
}
\description{
Run Diagnostic Plots for VASTWestCoast Output
}
\author{
Kelli Faye Johnson
}
