% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_EM.R
\name{VAST_EM}
\alias{VAST_EM}
\title{Run the full estimation method that loops over parameterizations}
\usage{
VAST_EM(reps, settings, directory, n_cluster, getdatafrom)
}
\arguments{
\item{reps}{A vector of replicate numbers}

\item{settings}{A list of settings.}

\item{directory}{The directory for the given simulation.}

\item{n_cluster}{An integer value specifying the number of cores to use.}

\item{getdatafrom}{The directory that stores the data.}
}
\value{
Saves information to the disk.
}
\description{
Run the full estimation method that loops over parameterizations
}
\author{
Kelli Faye Johnson
}
