% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_EM.R
\name{VAST_EMrepi}
\alias{VAST_EMrepi}
\title{Run the estimation method for a given set of "settings"}
\usage{
VAST_EMrepi(settings, data, datadir, emdir, overdispersion = c(eta1 = 0, eta2
  = 0))
}
\arguments{
\item{settings}{A list of settings.}

\item{data}{A matrix of data}

\item{datadir}{A directory where the kmeans information is housed.}

\item{emdir}{A directory where the EM information will be saved.}

\item{overdispersion}{A vector of values specifying if a vessel_year
effect should be included. The default \code{c("eta1" = 0, "eta2" = 0)}
does not estimate an effect.}
}
\value{
Saves information to the disk.
}
\description{
Run the estimation method for a given set of "settings"
}
\author{
Kelli Faye Johnson
}
