% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_mesh.R
\name{VAST_mesh}
\alias{VAST_mesh}
\title{Create grid and mesh for VASTWestCoast}
\usage{
VAST_mesh(
  data,
  survey,
  savedir = getwd(),
  numknots,
  range.depth = c(-35, -Inf),
  range.lat = c(-Inf, Inf)
)
}
\arguments{
\item{data}{A data frame extracted from the NWFSC data warehouse using
\link{get_data}. These points will be used as starting locations
for the knot locations in the final mesh. Locations do not need to be
positive tows, and all tow locations can be included, which is the default.}

\item{survey}{A character value supplying the survey name.
The value is typically an acronym.
See \code{\link[nwfscSurvey]{createMatrix}()} for a list of defaults.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{numknots}{The desired number of knots to have in the resulting mesh.
This will be approximate because the distance between points used for knots
will be used to solve to the ideal number of knots using a while loop.
Thus, expect the resulting number of knots to be within plus or minus five
percent of this input value. See \code{\link{get_mesh}()} for specifics.
The desired value is typically stored in the settings list; see,
\code{\link{get_settings}()} and is known as \code{"n_x"} in
\code{\link[FishStatsUtils]{make_settings}()}.}

\item{range.depth}{A vector of two numeric values specifying the range of
depth values of interest. Values must be in meters and negative values
represent those below sea level and positive values are assumed to be
above sea level. The first entry is the shallow value and the second
entry is the deeper cutoff. The default value is
\code{c(-35, -Inf)}, which truncates the data to depths deeper than 35 m
with no limit on how deep they can be.}

\item{range.lat}{A vector of two numeric values specifying the range of
latitude values of interest. Values must be in decimal degrees.
The first entry is the smaller value, i.e., southern border, and the second
entry is the larger value, i.e., northern border. The default value is
\code{c(-Inf, Inf)}, which leads to a range of the entire coast and if the
input argument is being used to truncate the data, the data is not changed.
To limit the data to California use \code{c(32.5, 42)} and
to limit the data to Washington and Oregon use \code{c(42, 49)}.
These ranges typically match up with the outer ranges of all
combined strata you wish to predict to, but
this does not have to be the case.}
}
\value{
A list with two elements is returned, the first is a input grid
for \code{"extrapolation_args"} and the second is an anisotropic
INLA triangulation mesh for \code{"spatial_args"}. Each of these list
elements are also plotted and the plots are saved in \code{savedir}.
}
\description{
Create the input grid and the mesh based on the survey and
format all objects as needed for VASTWestCoast.
}
\seealso{
See \link{get_inputgrid}, \link{get_mesh},
\link{plot.inputgrid}, and \link{plot.mesh}.
}
\author{
Kelli Faye Johnson
}
