% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_OM.R
\name{VAST_OM}
\alias{VAST_OM}
\title{Run the operating model (OM) for a simulation.}
\usage{
VAST_OM(reps, dir = NULL, conditioning, ncluster = 2)
}
\arguments{
\item{reps}{A vector of replicates you want to run in the OM folder.}

\item{dir}{A directory to place the simulated replicates. If specified
as \code{NULL}, which is the default, then the OM folder will be placed
in the conditioning directory and labeled sequentially with two numerical
integers followed by the letters OM. For example, \code{"02OM"} if there is
already an OM folder in the conditioning directory.}

\item{conditioning}{A list of conditioning specifications.}

\item{ncluster}{A integer value specifying the number of cores to use when
generating the simulated data.}
}
\value{
Nothing is returned from the function. Objects are saved
to the disk within the \code{\link{VAST_OMrepi}} function.
}
\description{
Run the operating model (OM) for a simulation.
}
\author{
Kelli Faye Johnson
}
