% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_OM.R
\name{VAST_OMrepi}
\alias{VAST_OMrepi}
\title{Run an individual replicate for the OM}
\usage{
VAST_OMrepi(omdir, rep, settings, maxtries = 10, plotindex = TRUE)
}
\arguments{
\item{omdir}{A file path where the \code{rep} should be saved, which will
lead to \code{file.path(omdir, rep)}.}

\item{rep}{A single numerical value specifying the replicate number.}

\item{settings}{A list of settings for the simulation.}

\item{maxtries}{The maximum number of tries to find a data set that does not
have all positive or zero proportions in a given year. A unique random number
is used per try.}

\item{plotindex}{A logical value specifying if the time-series trajectory
of the simulated data should be plotted in the \code{omdir}.}
}
\value{
Objects are saved in an \code{RData} file within the directory created
for the replicate. Also a plot is saved if \code{plotindex == TRUE} of the
time-series trajectory.
}
\description{
Run an individual replicate for the OM
}
\author{
Kelli Faye Johnson
}
