% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_run.R
\name{VAST_run}
\alias{VAST_run}
\title{Estimate parameters from a spatiotemporal index-standardization model}
\usage{
VAST_run(
  datalist,
  depth = c("no", "linear", "squared"),
  overdispersion,
  obsmodel,
  rundir,
  Version = NULL,
  calcs = c(1, rep(0, 6), 0, 1),
  strata,
  pass = FALSE,
  savefile,
  field = NULL,
  rho = NULL,
  comp = FALSE
)
}
\arguments{
\item{datalist}{A list of data used for the model.}

\item{depth}{A character value of either no, linear, or squared specifying how
the covariate should be included in the model.}

\item{overdispersion}{A vector of length two that specifies how overdispersion
is modelled. Typical options include \code{c("eta1"=0, "eta2"=0)} and
\code{c("Delta1"=1, "Delta2"=1)}, where the latter is for the US West Coast
and the former is for Alaska.}

\item{obsmodel}{A numerical vector of length two that specifies the distributions
for the observation and rate models. No default is specified, but the typical value
used on the US West Coast is \code{c(2, 0)}.}

\item{rundir}{A directory where the \code{.cpp} file will be compiled to.}

\item{Version}{The version of the \code{.cpp} file to use.}

\item{calcs}{A vector of length one or nine specifying which calculations
to perform. If a single integer of zero or one is given then this value will be used
for all eight calculations. Zero turns them off and one turns them on.}

\item{strata}{Strata specifications for calculating the index of abundance.}

\item{pass}{A logical value specifying if Pass should be include to estimate
catchability in the model. The default is to not use pass, i.e., \code{pass = FALSE}.}

\item{savefile}{The file name that you want the results saved to. A full path can
be specified, else the \code{.RData} file will be saved in the current directory.}

\item{field}{A vector of length four or one, where the latter situation 
will result in the value being repeated four times, 
to specify the number of factors or structure for spatial and spatiotemporal
effects for each model in turn, i.e.,
Omega1, Epsilon1, Omega2, and Epsilon2. The vector need not be named because
the names will be assigned in that order.
The default is a single value of \code{NULL} that results in searching through
the data for the appropriate configuration.
todo: Place more details in 'Details' section.}

\item{rho}{A vector of length four or one, where the latter situation 
will result in the value being repeated four times, 
to specify whether either intercepts (Beta1 and Beta2) and 
spatio-temporal variation (Epsilon1 and Epsilon2) are structured among 
time intervals. Options include \code{0:4}, where
if 0, then each year is modelled as a fixed effect; 
if 1, then each year is modelled as random following IID distribution; 
if 2, then each year is modelled as random following a random walk; 
if 3, then constant among years as fixed effect; and
if 4, then each year is modelled as random following an AR1 process).
todo: Place more details in 'Details' section.}

\item{comp}{A logical value specifying if the data are composition data.
The default is \code{FALSE}, where the data are assumed to be catch-rate data
used for an index of abundance rather than composition data.}
}
\value{
Nothing is returned directly, instead files are saved to the disk.
}
\description{
Estimate parameters from a spatiotemporal index-standardization model
}
\author{
Kelli Faye Johnson
}
