% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_setup.R
\name{VAST_setup}
\alias{VAST_setup}
\title{Set up the spatial information for a \link{[VAST]} run.}
\usage{
VAST_setup(data, dir, regionacronym, strata, nknots)
}
\arguments{
\item{dir}{A directory where you want to save the kmeans information to. The
character value can be a full path or relative to your current working directory.}
}
\description{
Set up the spatial information for a \link{[VAST]} run.
}
