% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_simulation.R
\name{VAST_simulation}
\alias{VAST_simulation}
\title{Run a complete simulation using \link{VAST}}
\usage{
VAST_simulation(maindir = getwd(), conditiondir = NULL, globalsettings,
  n_cluster)
}
\arguments{
\item{maindir}{The main directory that will house all of the results using a
specific folder structure, where the conditioning, operating model, and estimation
model results will all be housed in this directory.}

\item{conditiondir}{A directory for the conditioning. Only specify if the
conditioning has already been ran or you want to name the directory.}

\item{globalsettings}{A list of settings for the simulation.}

\item{n_cluster}{The number of clusters you want to use for parallel computing.}
}
\value{

}
\description{
Run a complete simulation using \link{VAST}
}
\examples{
VAST_simulation(globalsettings = Sim_Settings, n_cluster = 1)

}
\author{
Kelli Faye Johnson
}
