% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_spp.R
\name{VAST_spp}
\alias{VAST_spp}
\title{Run VAST Models for a Species}
\usage{
VAST_spp(dir, species, dist = c("lognormal", "gamma"))
}
\arguments{
\item{dir}{The directory where you want a folder named \code{species}
to be saved. The resulting folder will house all of the results from
VASTWestCoast in a directory called data.}

\item{species}{A common name for the given species of interest.}

\item{dist}{A vector of distributional assumptions that you want to include
for each species. Options are listed in the default call. Please contact
the package maintainer if you wish to use other distributions because names
other than those listed will not work.}
}
\value{
Nothing is returned from this function. Called functions save
information (e.g., plots and RData files) to the disk for viewing, further
research, or inclusion in reports.
}
\description{
Run VAST Models for a Species
}
\details{
\itemize{
\item Find surveys of interest in the data warehouse.
\item Find the strata for the species.
\item Run VAST_condition for each survey x distribution combo.
}
}
\author{
Kelli Faye Johnson
}
