% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_wc_diag_plot.R
\name{vast_wc_diag_plot}
\alias{vast_wc_diag_plot}
\title{Complete general diagnostic for West Coast VAST run.}
\usage{
vast_wc_diag_plot(data, rep, TmbData, Opt, dir = getwd(), Method = c("Mesh",
  "Grid", "Spherical_mesh"), grid_size_km = 25, randomseed = 1,
  nstart = 100, iter.max = 1000)
}
\arguments{
\item{data}{The raw data, though special column names are needed.}

\item{rep}{The Report as generated by code{Obj$report()}}

\item{TmbData}{The data used to fit the model and generated by
\code{\link[VAST]{Data_Fn}}}

\item{Opt}{Output from \code{\link[TMBhelper]{Optimize}}}

\item{dir}{The directory that you want to save the results to.
Do not include a trailing slash. The default is your current directory.}

\item{Method}{The method defining how space was represented in the model.
Options include \code{c("Mesh", "Grid", "Spherical_mesh")}.}

\item{grid_size_km}{The grid size used if \code{Method = "Grid"}.}

\item{randomseed}{Control value for the K-means algorithm
when \code{Methods = "Mesh"}.}

\item{nstart}{Control value for the K-means algorithm
when \code{Methods = "Mesh"}.}

\item{iter.max}{Control value for the K-means algorithm
when \code{Methods = "Mesh"}.}
}
\description{
Output figures and tables to the disk specific to a given index
of abundance generated by \code{\link[VAST]}. The code is specific
to data collected by the Northwest Fisheries Science Center
Bottom Trawl Survey.
}
\author{
Kelli Faye Johnson
}
