% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCGOP_clean.R
\name{WCGOP_clean}
\alias{WCGOP_clean}
\title{West Coast Groundfish Observer Program data preparation}
\usage{
WCGOP_clean(data, species, gear)
}
\arguments{
\item{data}{A data frame of WCOP data.}

\item{species}{A vector of character entries that specify the
species, using scientific names, that you want to keep.}

\item{gear}{A vector of character entries that specify the gear
types you want to keep. Vector can include values from both the
gear column and the GEAR_TYPE column.}
}
\value{
A n invisible data frame with the following columns
\enumerate{
  \item Catch_KG
  \item Retained_KG
  \item Discarded_KG
  \item Year
  \item Vessel
  \item Lat
  \item Lon
  \item Species
}
}
\description{
Prepare the data coming from the West Coast Groundfish Observer
Program (WCGOP) to be used in a \link{VAST} model.
Function was written as an example
}
\author{
Kelli Faye Johnson
}
