% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_unload.R
\name{clean_unload}
\alias{clean_unload}
\title{Unload VAST dll}
\usage{
clean_unload(searchfor = "VAST", keep = "VASTisCOOL")
}
\arguments{
\item{searchfor}{A file name of the dll to search for.
The files that are found will be removed from the loaded libraries.}

\item{keep}{A file name (or partial name) of the dll to search
for that you want to keep. For example, multiple libraries may
be loaded for VAST but you might only want to keep a certain one.
Put that certain one here, or some pattern that will match in a
regular expression search.}
}
\value{
The name(s) of the unloaded libraries are returned as
the first element of the list and the second element is all
dlls that were found.
}
\description{
Unload a dll library with the VAST name.
}
\author{
Kelli Faye Johnson
}
