% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage.R
\name{get_coverage}
\alias{get_coverage}
\title{Calculate coverage of an index}
\usage{
get_coverage(data, plot = FALSE)
}
\arguments{
\item{data}{A data frame with three x columns.}

\item{plot}{A logical value specifying whether or not to
plot the data}
}
\value{
A vector of information regarding the coverage for the data
that was provided in the argument \code{data}. One value will be returned
for each row of data.
}
\description{
Calculate coverage of an index
}
\author{
Kelli Faye Johnson
}
