% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_settings.R
\name{get_settings}
\alias{get_settings}
\title{Check that all of the settings are in the list}
\usage{
get_settings(settings = NULL, verbose = FALSE)
}
\arguments{
\item{settings}{A list of the current settings where each object in the
list must be named. Those names that are not found in the stored list will
be added. The default value of \code{NULL} leads to a full list being returned.}

\item{verbose}{A logical value specifying if the message should be output
to the screen or not.}
}
\value{
A list of setting for running a conditioning model or a simulation
with \code{\link{VAST}}.
}
\description{
Check that all of the settings are in the list
}
\examples{
get_settings(list("SigmaM" = 3))
names(get_settings(c("yes" = 2), verbose = TRUE))
}
\author{
Kelli Faye Johnson
}
