% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_condition.R
\name{VAST_condition}
\alias{VAST_condition}
\title{Condition an operating model based on fits to empirical data}
\usage{
VAST_condition(conditiondir, settings, spp, datadir, overdisperion = NULL,
  data = NULL)
}
\arguments{
\item{conditiondir}{A \code{file.path}, either full or relative, that will be used
to save the results. The directory will be created if it doesn't already exist.}

\item{settings}{A list of settings used to run the spatiotemporal model.}

\item{spp}{A character value that includes the survey acronym and the species name
in latin with all words separated by an underscore.
For example, \code{"WCGBTS_Sebastes_crameri"}.}

\item{datadir}{The directory where the database is stored.
Also, the kmean information specific to the given survey
used to collect the data will be saved inside \code{datadir} in a folder named
the same as the leading characters of \code{spp}. For example, \code{"WCGBTS"} for a
\code{spp = "WCGBTS_Sebastes_crameri"}, leading to
\code{file.path(datadir, "WCGBTS_Sebastes_crameri")}. The character value for the
directory should not have a trailing separator.}

\item{overdisperion}{A vector of overdispersion parameters to include
in the conditioning. The default is \code{NULL} and will enable the determination
of overdispersion from the \code{spp} based on its survey. If the survey occurs in
Alasksa (i.e., EBSBTS) then overdispersion will be based on the default value in
\code{\link[VAST]{Data_Fn}}of \code{c("eta1" = 0, "eta2" = 0)}. If the survey occurs
of the US West Coast, then overdispersion will be modelled for both components using a
vessel-year effect labeled delta.}
}
\value{
Nothing is returned by the function, but the function saves two \code{.RData}
structures to the disk in the code{conditiondir}.
}
\description{
Condition an operating model based on fits to empirical data
}
\author{
Kelli Faye Johnson
}
