% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_condition.R
\name{VAST_condition}
\alias{VAST_condition}
\title{Estimate Parameters by Fitting VAST to Empirical Data}
\usage{
VAST_condition(
  conditiondir,
  settings,
  spp,
  overdispersion,
  data = NULL,
  sensitivity = TRUE,
  compiledir = conditiondir
)
}
\arguments{
\item{conditiondir}{A directory, either a full or relative path,
that will be used to save the results.
The directory will be created recursively if it does not already exist.}

\item{settings}{A list of settings used to run the spatiotemporal model. The full
list of necessary settings can be seen by running \code{\link{get_settings}()}, and
any settings that are not included in the list supplied to this argument
will be added at their default values.}

\item{spp}{A character value that includes the survey acronym and the species name
in latin with all words separated by an underscore.
For example, \code{"WCGBTS_Sebastes_crameri"}.}

\item{overdispersion}{A vector of two integer values named "eta1" and "eta2"
that specify how overdispersion is modeled.
\code{"eta1"} is for the encounter model and
\code{"eta2"} is for the positive catch-rate model.
Typical options include
\code{c("eta1"=0, "eta2"=0)}, which is used for the
Alaska Fisheries Science Center Slope Survey, and
\code{c("eta1"=1, "eta2"=1)}, which is used for the
Northwest Fisheries Science Center
West Coast Groundfish Bottom Trawl Survey.
This vector is passed to \code{"OverdispersionConfig"} within the
\code{\link[VAST]{make_data}()} function.
The former configuration of zero turns off this feature in VAST.
The latter configuration allows for random covariation in catchability
(e.g., a vessel effect if v_i is assigned to vessel or a
vessel-year effect if v_i is assigned to a vector that combines vessel
and year into a factor). Specifically,
one random effect will be estimated for each unique value in v_i
if this feature is set to one.
The eta values can also be set to \code{"AR1"} if you want correlated
overdispersion in the available categories. The autoregressive process
might be appropriate if you believe there is overdispersion at the tow
level and tows closer in time provide more information about other tows than
tows that are further apart in time.
L_1 and L_2 are the estimated parameters representing the variance of these
random effects for each model.
The configuration for the overdispersion parameters is automatically set in
\code{\link{VAST_condition}} based on the input argument \code{spp},
which includes a short-hand version of the survey name. Leaving the input
value to the default of \code{NULL} allows the overdispersion parameter to
be set behind the scenes. This can be overridden by providing a vector.
If there is no default value for overdispersion, then a named vector
must be supplied where the first value is for the encounter model and the
second value is for the positive catch-rate model as given in the examples
above for the different surveys.}

\item{data}{A data frame that can be passed to the conditioning function
such that no data will be downloaded. todo: document the columns that are needed.}

\item{sensitivity}{logical; run sensitivity analyses specific to a given survey.
For example, the Triennial survey can be split into two separate surveys and limited
to 366 m depth.}

\item{compiledir}{A file path, either relative or full, to a directory where
the VAST cpp file will be compiled leading to three files saved to the disk.
Thus, users should have both read and write capabilities within this
directory. A single file path can be used for multiple models allowing for
the model to be only compiled once rather than multiple times.
This file path is tranfered to the \code{CompileDir} within
\code{\link[VAST]{make_model}()}.}
}
\value{
Nothing is returned by the function, but the function saves two \code{.RData}
structures to the disk in the \code{conditiondir}.
}
\description{
Fits data loaded in your R session
or data that is downloaded from the 
Northwest Fisheries Science Center datawarehouse
to a \code{\link[VAST]{make_model}} using
\code{\link[TMBhelper]{fit_tmb}}.
For more details on how the VAST model
is configured see the sections below.
}
\section{Pass}{

The US West Coast has operated using two passes of the survey area, where
each pass is done during a different time of year. If you think that migration
happens temporally and this migration will affect spatial density, then you
should include pass as a covariate. The default is to ignore pass, but I
presume that many US West Coast users will want to change this using the
\code{settings} argument.
}

\examples{
\dontrun{
# Use the default settings to run the model for WCGBTS - sablefish
VAST_condition(conditiondir = getwd(), settings = get_settings(),
  spp = settings$Species, sensitivity = FALSE)
}

}
\seealso{
See \code{\link{get_settings}()} for a list of default settings. Anything not
included in your own list supplied to the settings argument will be taken from this
default list.
}
\author{
Kelli Faye Johnson
}
