% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_diagnostics.R
\name{VAST_diagnostics}
\alias{VAST_diagnostics}
\title{Run diagnostic plots for VASTWestCoast output}
\usage{
VAST_diagnostics(dir = getwd())
}
\arguments{
\item{dir}{The directory where the output from
\link{VAST_condition} is saved. Can be a relative
or an absolute path.}
}
\value{
Nothing is returned.
\itemize{
\item convergence_gradient.txt: Parameter estimates
\item VASTWestCoast_QQ.png: Quantile-Quantile plot
}
}
\description{
Run diagnostic plots for VASTWestCoast output
}
\author{
Kelli F. Johnson
}
