% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_do.R
\name{VAST_do}
\alias{VAST_do}
\title{Shortened Code for VAST}
\usage{
VAST_do(Database, settings, conditiondir)
}
\arguments{
\item{Database}{A data base ran through \code{\link{get_data}()}.}

\item{settings}{A list of settings used to run the spatiotemporal model. The full
list of necessary settings can be seen by running \code{\link{get_settings}()}, and
any settings that are not included in the list supplied to this argument
will be added at their default values.}

\item{conditiondir}{A directory, either full or relative, that will be used
to save the results. The directory will be created if it doesn't already exist.}
}
\description{
Major workhorse to run a VAST model
}
