% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAST_simulation.R
\name{VAST_simulation}
\alias{VAST_simulation}
\title{Run a complete simulation using \link{VAST}}
\usage{
VAST_simulation(maindir = getwd(), conditiondir = NULL, globalsettings,
  n_cluster, OMdir = NULL)
}
\arguments{
\item{maindir}{The main directory that will house all of the results using a
specific folder structure, where the conditioning, operating model, and estimation
model results will all be housed in this directory.}

\item{conditiondir}{A directory for the conditioning. Only specify if the
conditioning has already been ran or you want to name the directory.}

\item{globalsettings}{A list of settings for the simulation.
The full settings need not be supplied. Any settings not included in the list will
be added using \code{\link{get_settings}}.}

\item{n_cluster}{The number of clusters you want to use for parallel computing.}

\item{OMdir}{A relative path from the conditiondir for the directory for the OM,
so that you can run more iterations if you so wish.}
}
\value{
A data frame of results are returned that contains
information about the operating models and estimation methods.
}
\description{
Run a complete simulation using \link{VAST}
}
\examples{
\dontrun{
VAST_simulation(
  globalsettings = list("Species" = "WCGBTS_Anoplopoma_fimbria"), 
  n_cluster = 1)
}

}
\author{
Kelli Faye Johnson
}
